We invent a learn way for units.
Units have a learn method, that has a pointer to LearnAlgorithm object as a argument.
LearnAlgorithm is a basic class for algorithms which work with one-leyer units
(not a multiple-input). To use some different learn-algorithm you need to
override start-method and send learnAlg-object to argument of learn method of unit.